:sparkles: Added components to test LBP classifier on equalized dataset

+++ matlab_dir/equalize_dataset.m
@@ -1,10 +1,14 @@
 function equalize_dataset(images)
     folder_name = 'equalized_dataset';
-    mkdir(folder_name);
-    for i = 1:size(images)
-        label = strcat('dataset/', string(images(i)));
-        image = im2double(imread(label));
-        equalized = equalize_image(image);
-        imwrite(equalized, strcat(folder_name, '/', string(images(i))));
+    if ~exist(folder_name, 'dir')
+        mkdir(folder_name);
+        for i = 1:size(images)
+            label = strcat('dataset/', string(images(i)));
+            image = im2double(imread(label));
+            equalized = equalize_image(image);
+            imwrite(equalized, strcat(folder_name, '/', string(images(i))));
+        end
+    else
+        disp("Equalized dataset already exists");
     end
 end
